dev: created base connection pool
