+++ config/go.mod
@@ -2,4 +2,4 @@ module config
 
 go 1.24.4
 
-require github.com/joho/godotenv v1.5.1 // indirect
+require github.com/joho/godotenv v1.5.1
