+++ types/config/config.go
@@ -1,0 +1,8 @@
+package types
+
+type PsConfig struct {
+	Host     string
+	Username string
+	Password string
+	Db       string
+}
