+++ types/database/users/users_record.go
@@ -1,0 +1,11 @@
+package types
+
+import "time"
+
+type UsersRecord struct {
+	Id        int
+	Login     string
+	Password  string
+	CreatedAt time.Time
+	UpdatedAt time.Time
+}
