+++ types/go.mod
@@ -1,0 +1,3 @@
+module types
+
+go 1.24.4
